Memoクラスを別ファイルに切り出した
他のnpmもES Module syntaxで読み込むように変更した
