+++ 04.class/memo.js
@@ -1,51 +1,11 @@
 #! /usr/bin/env node
-const { v4: uuidv4 } = require("uuid");
-const fs = require("node:fs/promises");
-const argv = require("minimist")(process.argv.slice(2));
-const { Select } = require("enquirer");
 
-class Memo {
-  constructor({ id = null, body }) {
-    this.id = this.#getId(id);
-    this.body = body;
-  }
-
-  static async of(id) {
-    const file = await fs.readFile(`database/${id}.json`);
-    const memo_json = await JSON.parse(file);
-    const memo_class = await new Memo({
-      id: memo_json.id,
-      body: memo_json.body,
-    });
-    return memo_class;
-  }
-
-  #getId(id) {
-    if (id === null) {
-      return uuidv4();
-    } else {
-      return id;
-    }
-  }
-
-  getFirstLine() {
-    return this.body.split(/\n/)[0];
-  }
-
-  getFileName() {
-    return `${this.id}.json`;
-  }
-
-  createFile() {
-    fs.writeFile(`database/${this.getFileName()}`, JSON.stringify(this))
-      .then(() => {
-        console.log("正常に書き込みが完了しました");
-      })
-      .catch((err) => {
-        console.error(err);
-      });
-  }
-}
+import * as fs from "node:fs/promises";
+import pkg from "enquirer";
+import minimist from "minimist";
+import { Memo } from "./MemoModel.js";
+const { Select } = pkg;
+const argv = minimist(process.argv.slice(2));
 
 async function listFirstLines() {
   (await fetchAllMemos()).forEach((memo) => {
@@ -122,8 +82,8 @@ if (argv.l) {
     .then((value) => {
       return new Memo({ body: value });
     })
-    .then((memo) => {
-      memo.createFile();
+    .then((memo_obj) => {
+      memo_obj.createFile();
     })
     .catch((err) => {
       console.error(err);
